Display tooltips for unresolved links
ScalaCommentToContentConverter needed to be refactored to inherit from
DocTagToContentConverter (because our version of `buildContent` needs to always
be called first before the super version, otherwise nested tags aren't seen by
our implementation). Since the latter is an object, its code had to be copied to
an open class instead (which is OpenDocTagToContentConverter).